requirements.txt
asgiref==3.9.1
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
idna==3.10
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
psycopg2-binary==2.9.10
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
